requirements.txt
cloudpickle==3.1.1
contourpy==1.3.0
cycler==0.12.1
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
gym==0.26.2
gym-notices==0.0.8
importlib_metadata==8.6.1
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.2.1
numpy==1.24.3
packaging==24.2
pillow==11.2.1
pygame==2.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0
six==1.17.0
sympy==1.13.1
torch==2.0.0
typing_extensions==4.13.2
zipp==3.21.0
